+++ Changelog.md
@@ -9,6 +9,8 @@
 #### :wrench: Fix
 
 1. Fix: Correct a typo of the html tags in the readme file.
+2. Fix: Fix typos in the docstrings.
+3. Fix: Make `caches.typehints.Deferred` used by other modules.
 
 #### :floppy_disk: Change